requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pillow==10.4.0
torch
transformers==4.51.3
tokenizers==0.21.1
accelerate
python-multipart
huggingface_hub
